added unbind_regs and modified dummy emulator to show how useful it is
the idea here is that we can get handles to the emulated registers in the same
way we create new JIT variables. Then we work with the emu reg handles and the
value magically makes its way back to the emulator register array. What really
happens is that the handles point to the emu reg's location in the JIT code's
stack. If this handle happens to be mapped to a native register at the end of
our function then the native register's value is written to the emu reg's
location in the stack. This happens after variables have been freed (i.e. the
emu regs are at the bottom of the stack) so this frees up all native registers
giving us flexibility. Then we write the emu reg values from the stack to the
emulator's register array using the first emulator pointer. While it's possible
to write the emu reg values directly from the native registers to the array,
this is much more complicated. After implementing this in the first few versions
of this module I decided the performance hit is probably worth it for now.