added labels and jump to Recompiler
Labels and jump are now usable, but I should fix a subtle bug related to
conditionally-executed JIT code. The problem is that stack offsets in a
conditionally-executed section that are balanced by offsets in
unconditionally-executed sections may cause stack overflows. Stack offsets in
conditionally-executed sections will either be balanced (e.g. conditionally
calling an external function is balanced by the function's return) or allocating
new variables. There are three options for handling new variables:

  (1) create them in unconditionally-executed sections
  (2) use new_variable to track necessary stack space then retroactively (i.e.
      right before calling compile()) emit an addq to allocate stack space at
      the beginning
  (3) reintroduce scopes that call free_variables when closed.

(2) sounds good, but introduces lots of dependencies (i.e. any time full_stack()
is called to determine some stack offset the value that is emitted will have to
be overwritten at the end after preallocating the stack space). (1) is doable
but burdens the user so that's off the table. (3) seems pretty good, but it
still is a bit burdensome for the user so (2) is probably the best long-term
strategy and (1) is better for the short-term