made JITValue::Flags act as a queue
note that x86 registers mapped to JITValue::Flags can only be unbound by
Allocator::load_flags() so in the case that there are 15 branch delay slots in a
row (e.g. 15 ADDI's in a row) this won't work. I could give JITValue::Flags a
spot on the stack, but it might be a bit tricky to use since
Recompiler::prepare_for_exit() cleans up the stack